jupyter/requirements.txt
jupyterlab
numpy
pandas[performance]
polars
scipy
statsmodels
matplotlib
seaborn
scikit-learn
requests
python-dotenv
psycopg
sqlalchemy
